+++ src/Composer/Satis/Command/BuildCommand.php
@@ -13,28 +13,15 @@
 namespace Composer\Satis\Command;
 
 use Composer\Config\JsonConfigSource;
-use Composer\Package\Loader\ArrayLoader;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Composer\Command\Command;
-use Composer\DependencyResolver\Pool;
 use Composer\Composer;
 use Composer\Config;
-use Composer\Package\Dumper\ArrayDumper;
-use Composer\Package\AliasPackage;
-use Composer\Package\BasePackage;
-use Composer\Package\LinkConstraint\MultiConstraint;
-use Composer\Package\PackageInterface;
-use Composer\Package\Link;
-use Composer\Repository\ComposerRepository;
-use Composer\Repository\PlatformRepository;
 use Composer\Json\JsonFile;
-use Composer\Factory;
-use Composer\Util\Filesystem;
 use Composer\Util\RemoteFilesystem;
-use Composer\IO\ConsoleIO;
 
 /**
  * @author Jordi Boggiano <j.boggiano@seld.be>
@@ -102,7 +89,7 @@ EOT
         $verbose = $input->getOption('verbose');
         $configFile = $input->getArgument('file');
         $packagesFilter = $input->getArgument('packages');
-        $skipErrors = (bool)$input->getOption('skip-errors');
+        $skipErrors = (bool) $input->getOption('skip-errors');
 
         // load auth.json authentication information and pass it to the io interface
         $io = $this->getIO();
@@ -135,7 +122,7 @@ EOT
             $requireAll = true;
         }
 
-        $minimumStability =  isset($config['minimum-stability']) ? $config['minimum-stability'] : 'dev';
+        $minimumStability = isset($config['minimum-stability']) ? $config['minimum-stability'] : 'dev';
 
         if (!$outputDir = $input->getArgument('output-dir')) {
             $outputDir = isset($config['output-dir']) ? $config['output-dir'] : null;
@@ -146,14 +133,14 @@ EOT
         }
 
         $composer = $this->getApplication()->getComposer(true, $config);
-        $packages = $this->selectPackages($composer, $output, $verbose, $requireAll, $requireDependencies, $requireDevDependencies, $minimumStability, $skipErrors, $packagesFilter);
-
-        if ($htmlView = !$input->getOption('no-html-output')) {
-            $htmlView = !isset($config['output-html']) || $config['output-html'];
-        }
+        #JamesRezo
+        $packagesList = new \Composer\Satis\Builder\PackagesList();
+        $packages = $packagesList->select($composer, $output, $verbose, $requireAll, $requireDependencies, $requireDevDependencies, $minimumStability, $skipErrors, $packagesFilter);
 
         if (isset($config['archive']['directory'])) {
-            $this->dumpDownloads($config, $packages, $input, $output, $outputDir, $skipErrors);
+            #JamesRezo
+            $downloads = new \Composer\Satis\Builder\Downloads();
+            $downloads->dump($config, $packages, $input, $output, $outputDir, $skipErrors);
         }
 
         $filenamePrefix = $outputDir.'/include/all';
@@ -161,19 +148,16 @@ EOT
         if (!empty($packagesFilter)) {
             // in case of an active package filter we need to load the dumped packages.json and merge the
             // updated packages in
-            $oldPackages = $this->loadDumpedPackages($filename, $packagesFilter);
+            $oldPackages = $packagesList->load($filename, $packagesFilter);
             $packages += $oldPackages;
             ksort($packages);
         }
 
-        $packageFile = $this->dumpPackageIncludeJson($packages, $output, $filenamePrefix);
-        $packageFileHash = hash_file('sha1', $packageFile);
+        $packagesList->dump($packages, $output, $filenamePrefix);
 
-        $includes = array(
-            'include/all$'.$packageFileHash.'.json' => array( 'sha1'=>$packageFileHash ),
-        );
-
-        $this->dumpPackagesJson($includes, $output, $filename);
+        if ($htmlView = !$input->getOption('no-html-output')) {
+            $htmlView = !isset($config['output-html']) || $config['output-html'];
+        }
 
         if ($htmlView) {
             $dependencies = array();
@@ -185,394 +169,10 @@ EOT
 
             $rootPackage = $composer->getPackage();
             $twigTemplate = isset($config['twig-template']) ? $config['twig-template'] : null;
-            $this->dumpWeb($packages, $output, $rootPackage, $outputDir, $twigTemplate, $dependencies);
+            #JamesRezo
+            $web = new \Composer\Satis\Builder\Web();
+            $web->dump($packages, $output, $rootPackage, $outputDir, $twigTemplate, $dependencies);
         }
-    }
-
-    private function selectPackages(Composer $composer, OutputInterface $output, $verbose, $requireAll, $requireDependencies, $requireDevDependencies, $minimumStability, $skipErrors, array $packagesFilter = array())
-    {
-        $selected = array();
-
-        // run over all packages and store matching ones
-        $output->writeln('<info>Scanning packages</info>');
-
-        $repos = $composer->getRepositoryManager()->getRepositories();
-        $pool = new Pool($minimumStability);
-        foreach ($repos as $repo) {
-            try {
-                $pool->addRepository($repo);
-            } catch (\Exception $exception) {
-                if (!$skipErrors) {
-                    throw $exception;
-                }
-                $output->writeln(sprintf("<error>Skipping Exception '%s'.</error>", $exception->getMessage()));
-            }
-        }
-
-        if ($requireAll) {
-            $links = array();
-            $filterForPackages = count($packagesFilter) > 0;
-
-            foreach ($repos as $repo) {
-                // collect links for composer repos with providers
-                if ($repo instanceof ComposerRepository && $repo->hasProviders()) {
-                    foreach ($repo->getProviderNames() as $name) {
-                        $links[] = new Link('__root__', $name, new MultiConstraint(array()), 'requires', '*');
-                    }
-                } else {
-                    $packages = array();
-                    if ($filterForPackages) {
-                        // apply package filter if defined
-                        foreach ($packagesFilter as $filter) {
-                            $packages += $repo->findPackages($filter);
-                        }
-                    } else {
-                        // process other repos directly
-                        $packages = $repo->getPackages();
-                    }
-
-                    foreach ($packages as $package) {
-                        // skip aliases
-                        if ($package instanceof AliasPackage) {
-                            continue;
-                        }
-
-                        if ($package->getStability() > BasePackage::$stabilities[$minimumStability]) {
-                            continue;
-                        }
-
-                        // add matching package if not yet selected
-                        if (!isset($selected[$package->getUniqueName()])) {
-                            if ($verbose) {
-                                $output->writeln('Selected '.$package->getPrettyName().' ('.$package->getPrettyVersion().')');
-                            }
-                            $selected[$package->getUniqueName()] = $package;
-                        }
-                    }
-                }
-            }
-        } else {
-            $links = array_values($composer->getPackage()->getRequires());
-
-            // only pick up packages in our filter, if a filter has been set.
-            if (count($packagesFilter) > 0) {
-                 $links = array_filter($links, function(Link $link) use ($packagesFilter) {
-                     return in_array($link->getTarget(), $packagesFilter);
-                });
-            }
-
-            $links = array_values($links);
-        }
-
-
-        // process links if any
-        $depsLinks = array();
-
-        $i = 0;
-        while (isset($links[$i])) {
-            $link = $links[$i];
-            $i++;
-            $name = $link->getTarget();
-            $matches = $pool->whatProvides($name, $link->getConstraint(), true);
-
-            foreach ($matches as $index => $package) {
-                // skip aliases
-                if ($package instanceof AliasPackage) {
-                    $package = $package->getAliasOf();
-                }
-
-                // add matching package if not yet selected
-                if (!isset($selected[$package->getUniqueName()])) {
-                    if ($verbose) {
-                        $output->writeln('Selected '.$package->getPrettyName().' ('.$package->getPrettyVersion().')');
-                    }
-                    $selected[$package->getUniqueName()] = $package;
-
-                    if (!$requireAll) {
-                        $required = array();
-                        if ($requireDependencies) {
-                            $required = $package->getRequires();
-                        }
-                        if ($requireDevDependencies) {
-                            $required = array_merge($required, $package->getDevRequires());
-                        }
-                        // append non-platform dependencies
-                        foreach ($required as $dependencyLink) {
-                            $target = $dependencyLink->getTarget();
-                            if (!preg_match(PlatformRepository::PLATFORM_PACKAGE_REGEX, $target)) {
-                                $linkId = $target.' '.$dependencyLink->getConstraint();
-                                // prevent loading multiple times the same link
-                                if (!isset($depsLinks[$linkId])) {
-                                    $links[] = $dependencyLink;
-                                    $depsLinks[$linkId] = true;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            if (!$matches) {
-                $output->writeln('<error>The '.$name.' '.$link->getPrettyConstraint().' requirement did not match any package</error>');
-            }
-        }
-
-        ksort($selected, SORT_STRING);
-
-        return $selected;
-    }
-
-    /**
-     * @param array           $config   Directory where to create the downloads in, prefix-url, etc..
-     * @param array           $packages
-     * @param InputInterface  $input
-     * @param OutputInterface $output
-     * @param string          $outputDir
-     * @param bool            $skipErrors   If true, any exception while dumping a package will be ignored.
-     *
-     * @return void
-     */
-    private function dumpDownloads(array $config, array $packages, InputInterface  $input, OutputInterface $output, $outputDir, $skipErrors)
-    {
-        if (isset($config['archive']['absolute-directory'])) {
-            $directory = $config['archive']['absolute-directory'];
-        } else {
-            $directory = sprintf('%s/%s', $outputDir, $config['archive']['directory']);
-        }
-
-        $output->writeln(sprintf("<info>Creating local downloads in '%s'</info>", $directory));
-
-        $format = isset($config['archive']['format']) ? $config['archive']['format'] : 'zip';
-        $endpoint = isset($config['archive']['prefix-url']) ? $config['archive']['prefix-url'] : $config['homepage'];
-        $skipDev = isset($config['archive']['skip-dev']) ? (bool) $config['archive']['skip-dev'] : false;
-        $whitelist = isset($config['archive']['whitelist']) ? (array) $config['archive']['whitelist'] : array();
-        $blacklist = isset($config['archive']['blacklist']) ? (array) $config['archive']['blacklist'] : array();
-
-        $includeArchiveChecksum = isset($config['archive']['checksum']) ? (bool) $config['archive']['checksum'] : true;
-
-        $composerConfig = Factory::createConfig();
-        $factory = new Factory;
-        $io = new ConsoleIO($input, $output, $this->getApplication()->getHelperSet());
-        $io->loadConfiguration($composerConfig);
-
-        /* @var \Composer\Downloader\DownloadManager $downloadManager */
-        $downloadManager = $factory->createDownloadManager($io, $composerConfig);
-
-        /* @var \Composer\Package\Archiver\ArchiveManager $archiveManager */
-        $archiveManager = $factory->createArchiveManager($composerConfig, $downloadManager);
-
-        $archiveManager->setOverwriteFiles(false);
-
-        shuffle($packages);
-        /* @var \Composer\Package\CompletePackage $package */
-        foreach ($packages as $package) {
-            if ('metapackage' === $package->getType()) {
-                continue;
-            }
-
-            $name = $package->getName();
-
-            if (true === $skipDev && true === $package->isDev()) {
-                $output->writeln(sprintf("<info>Skipping '%s' (is dev)</info>", $name));
-                continue;
-            }
-
-            $names = $package->getNames();
-            if ($whitelist && !array_intersect($whitelist, $names)) {
-                $output->writeln(sprintf("<info>Skipping '%s' (is not in whitelist)</info>", $name));
-                continue;
-            }
-
-            if ($blacklist && array_intersect($blacklist, $names)) {
-                $output->writeln(sprintf("<info>Skipping '%s' (is in blacklist)</info>", $name));
-                continue;
-            }
-
-            $output->writeln(sprintf("<info>Dumping '%s'.</info>", $name));
-
-            try {
-                if ('pear-library' === $package->getType()) {
-                    // PEAR packages are archives already
-                    $filesystem = new Filesystem();
-                    $packageName = $archiveManager->getPackageFilename($package);
-                    $path =
-                        realpath($directory) . '/' . $packageName . '.' .
-                        pathinfo($package->getDistUrl(), PATHINFO_EXTENSION);
-                    if (!file_exists($path)) {
-                        $downloadDir = sys_get_temp_dir() . '/composer_archiver/' . $packageName;
-                        $filesystem->ensureDirectoryExists($downloadDir);
-                        $downloadManager->download($package, $downloadDir, false);
-                        $filesystem->ensureDirectoryExists($directory);
-                        $filesystem->rename($downloadDir . '/' . pathinfo($package->getDistUrl(), PATHINFO_BASENAME), $path);
-                        $filesystem->removeDirectory($downloadDir);
-                    }
-                    // Set archive format to `file` to tell composer to download it as is
-                    $archiveFormat = 'file';
-                } else {
-                    $path = $archiveManager->archive($package, $format, $directory);
-                    $archiveFormat = $format;
-                }
-                $archive = basename($path);
-                $distUrl = sprintf('%s/%s/%s', $endpoint, $config['archive']['directory'], $archive);
-                $package->setDistType($archiveFormat);
-                $package->setDistUrl($distUrl);
-
-                if ($includeArchiveChecksum) {
-                    $package->setDistSha1Checksum(hash_file('sha1', $path));
-                }
-
-                $package->setDistReference($package->getSourceReference());
-            } catch (\Exception $exception) {
-                if (!$skipErrors) {
-                    throw $exception;
-                }
-                $output->writeln(sprintf("<error>Skipping Exception '%s'.</error>", $exception->getMessage()));
-            }
-        }
-    }
-
-
-    private function dumpPackageIncludeJson(array $packages, OutputInterface $output, $filename)
-    {
-        $repo = array('packages' => array());
-        $dumper = new ArrayDumper;
-        foreach ($packages as $package) {
-            $repo['packages'][$package->getName()][$package->getPrettyVersion()] = $dumper->dump($package);
-        }
-        $repoJson = new JsonFile($filename);
-        $repoJson->write($repo);
-        $hash = hash_file('sha1', $filename);
-        $filenameWithHash = $filename.'$'.$hash.'.json';
-        rename($filename, $filenameWithHash);
-        $output->writeln("<info>wrote packages json $filenameWithHash</info>");
-        return $filenameWithHash;
-    }
-
-    private function dumpPackagesJson($includes, OutputInterface $output, $filename){
-        $repo = array(
-            'packages'          => array(),
-            'includes'          => $includes,
-        );
-
-        $output->writeln('<info>Writing packages.json</info>');
-        $repoJson = new JsonFile($filename);
-        $repoJson->write($repo);
-    }
-
-    private function dumpWeb(array $packages, OutputInterface $output, PackageInterface $rootPackage, $directory, $template = null, array $dependencies = array())
-    {
-        $templateDir = $template ? pathinfo($template, PATHINFO_DIRNAME) : __DIR__.'/../../../../views';
-        $loader = new \Twig_Loader_Filesystem($templateDir);
-        $twig = new \Twig_Environment($loader);
-
-        $mappedPackages = $this->getMappedPackageList($packages);
-
-        $name = $rootPackage->getPrettyName();
-        if ($name === '__root__') {
-            $name = 'A';
-            $output->writeln('Define a "name" property in your json config to name the repository');
-        }
-
-        if (!$rootPackage->getHomepage()) {
-            $output->writeln('Define a "homepage" property in your json config to configure the repository URL');
-        }
-
-        $output->writeln('<info>Writing web view</info>');
-
-        $content = $twig->render($template ? pathinfo($template, PATHINFO_BASENAME) : 'index.html.twig', array(
-            'name'          => $name,
-            'url'           => $rootPackage->getHomepage(),
-            'description'   => $rootPackage->getDescription(),
-            'packages'      => $mappedPackages,
-            'dependencies'  => $dependencies,
-        ));
-
-        file_put_contents($directory.'/index.html', $content);
-    }
-
-    private function loadDumpedPackages($filename, array $packagesFilter = array())
-    {
-        $packages = array();
-        $repoJson = new JsonFile($filename);
-        $dirName  = dirname($filename);
-
-        if ($repoJson->exists()) {
-            $loader       = new ArrayLoader();
-            $jsonIncludes = $repoJson->read();
-            $jsonIncludes = isset($jsonIncludes['includes']) && is_array($jsonIncludes['includes'])
-                ? $jsonIncludes['includes']
-                : array();
-
-            foreach ($jsonIncludes as $includeFile => $includeConfig) {
-                $includeJson = new JsonFile($dirName . '/' . $includeFile);
-                $jsonPackages = $includeJson->read();
-                $jsonPackages = isset($jsonPackages['packages']) && is_array($jsonPackages['packages'])
-                    ? $jsonPackages['packages']
-                    : array();
-
-                foreach ($jsonPackages as $jsonPackage) {
-                    if (is_array($jsonPackage)) {
-                        foreach ($jsonPackage as $jsonVersion) {
-                            if (is_array($jsonVersion)) {
-                                if (isset($jsonVersion['name']) && in_array($jsonVersion['name'], $packagesFilter)) {
-                                    continue;
-                                }
-                                $package = $loader->load($jsonVersion);
-                                $packages[$package->getUniqueName()] = $package;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        return $packages;
-    }
-
-    private function getMappedPackageList(array $packages)
-    {
-        $groupedPackages = $this->groupPackagesByName($packages);
-
-        $mappedPackages = array();
-        foreach ($groupedPackages as $name => $packages) {
-            $mappedPackages[$name] = array(
-                'highest' => $this->getHighestVersion($packages),
-                'versions' => $this->getDescSortedVersions($packages),
-            );
-        }
-
-        return $mappedPackages;
-    }
-
-    private function groupPackagesByName(array $packages)
-    {
-        $groupedPackages = array();
-        foreach ($packages as $package) {
-            $groupedPackages[$package->getName()][] = $package;
-        }
-
-        return $groupedPackages;
-    }
-
-    private function getHighestVersion(array $packages)
-    {
-        $highestVersion = null;
-        foreach ($packages as $package) {
-            if (null === $highestVersion || version_compare($package->getVersion(), $highestVersion->getVersion(), '>=')) {
-                $highestVersion = $package;
-            }
-        }
-
-        return $highestVersion;
-    }
-
-    private function getDescSortedVersions(array $packages)
-    {
-        usort($packages, function ($a, $b) {
-            return version_compare($b->getVersion(), $a->getVersion());
-        });
-
-        return $packages;
     }
 
     /**
@@ -591,11 +191,13 @@ EOT
             $config->merge(array('config' => $file->read()));
         }
         $config->setAuthConfigSource(new JsonConfigSource($file, true));
+
         return $config;
     }
 
     /**
      * @return string
+     *
      * @throws \RuntimeException
      */
     private function getComposerHome()
@@ -606,12 +208,12 @@ EOT
                 if (!getenv('APPDATA')) {
                     throw new \RuntimeException('The APPDATA or COMPOSER_HOME environment variable must be set for composer to run correctly');
                 }
-                $home = strtr(getenv('APPDATA'), '\\', '/') . '/Composer';
+                $home = strtr(getenv('APPDATA'), '\\', '/').'/Composer';
             } else {
                 if (!getenv('HOME')) {
                     throw new \RuntimeException('The HOME or COMPOSER_HOME environment variable must be set for composer to run correctly');
                 }
-                $home = rtrim(getenv('HOME'), '/') . '/.composer';
+                $home = rtrim(getenv('HOME'), '/').'/.composer';
             }
         }
